+++ api-js/src/mutations/organizations/__tests__/create-organization.test.js
@@ -4,20 +4,20 @@ const { ArangoTools, dbNameFromFile } = require('arango-tools')
 const { graphql, GraphQLSchema, GraphQLError } = require('graphql')
 const { toGlobalId } = require('graphql-relay')
 const { setupI18n } = require('@lingui/core')
-const { makeMigrations } = require('../../migrations')
-const { createQuerySchema } = require('../queries')
-const { createMutationSchema } = require('../mutations')
-const englishMessages = require('../locale/en/messages')
-const frenchMessages = require('../locale/fr/messages')
+const { makeMigrations } = require('../../../../migrations')
+const { createQuerySchema } = require('../../../queries')
+const { createMutationSchema } = require('../..')
+const englishMessages = require('../../../locale/en/messages')
+const frenchMessages = require('../../../locale/fr/messages')
 
 const bcrypt = require('bcrypt')
-const { cleanseInput, slugify } = require('../validators')
-const { tokenize, userRequired } = require('../auth')
+const { cleanseInput, slugify } = require('../../../validators')
+const { tokenize, userRequired } = require('../../../auth')
 const {
   orgLoaderBySlug,
   userLoaderByKey,
   userLoaderByUserName,
-} = require('../loaders')
+} = require('../../../loaders')
 
 const { DB_PASS: rootPass, DB_URL: url } = process.env
